Add DGDR conversion webhook and split conversion.go per type
Add v1alpha1 ↔ v1beta1 conversion for DynamoGraphDeploymentRequest
(DGDR) and split the monolithic conversion.go into per-type files
matching the existing *_types.go naming convention.

Hub marker: add DynamoGraphDeploymentRequest to v1beta1/hub.go.

DGDR conversion handles three categories of field mappings:
- Simple renames (Model, Backend, AutoApply, UseMocker, WorkersImage)
- JSON blob ↔ structured fields (SLA ttft/itl, Workload isl/osl,
  ModelCache) with full-blob annotation for round-trip preservation
- Annotation-based storage for v1alpha1 fields with no v1beta1
  equivalent (ProfilerImage, ConfigMapRef, OutputPVC, etc.)

State ↔ Phase mapping accounts for v1alpha1 "Ready" mapping to either
Ready or Deployed depending on Deployment.Created context.
